add agenda for may 7 workshop
